refactors to smaller React components now that projects has grown
